HR app

+++ HrApp/Model/EncashmentLeaveModel.cs
@@ -6,11 +6,11 @@ namespace HrApp.Model
         public int StuffId { get; set; }
         public string Name { get; set; }
         public DateTime Date { get; set; }
-        public int LeaveBalance { get; set; }
+        public double LeaveBalance { get; set; }
         public string StuffFirstNameLastName { get; set; }
         public int APPROVED_Encashment_Days { get; set; }
         public string APPROVED_Comments { get; set; }
-        public int balance { get; set; }
+        public double balance { get; set; }
         public int Encashment_Days { get; set; }
 
         public string Notes { get; set; }
